+++ intercom_flutter/ios/Classes/IntercomFlutterPlugin.m
@@ -86,7 +86,7 @@ id unread;
         result(@"Presented messenger");
     }
     else if([@"hideMessenger" isEqualToString:call.method]) {
-        [Intercom hideMessenger];
+        [Intercom hideIntercom];
         result(@"Messenger hidden");
     }
     else if([@"displayHelpCenter" isEqualToString:call.method]) {
@@ -125,6 +125,19 @@ id unread;
             NSData* encodedToken=[token dataUsingEncoding:NSUTF8StringEncoding];
             [Intercom setDeviceToken:encodedToken];
             result(@"Token set");
+        }
+    } else if([@"displayArticle" isEqualToString:call.method]) {
+        NSString *articleId = call.arguments[@"articleId"];
+        NSLog(@"%@", articleId);
+        if(articleId != (id)[NSNull null] && articleId != nil) {
+            [Intercom presentArticle:articleId];
+            result(@"displaying article");
+        }
+    } else if([@"displayCarousel" isEqualToString:call.method]) {
+        NSString *carouselId = call.arguments[@"carouselId"];
+        if(carouselId != (id)[NSNull null] && carouselId != nil) {
+            [Intercom presentCarousel:carouselId];
+            result(@"displaying carousel");
         }
     }
     else {