+++ intercom_flutter/lib/intercom_flutter.dart
@@ -128,4 +128,27 @@ class Intercom {
   static Future<void> handlePush(Map<String, dynamic> message) async {
     return IntercomFlutterPlatform.instance.handlePush(message);
   }
+
+  /// To display an Article, pass in an [articleId] from your Intercom workspace.
+  ///
+  /// An article must be ‘live’ to be used in this feature.
+  /// If it is in a draft or paused state,
+  /// end-users will see an error if the app tries to open the content.
+  ///
+  /// known issue:
+  /// If the articles feature is not enabled on Intercom account
+  /// then opening the article will crash the app on iOS.
+  /// see https://forum.intercom.com/s/question/0D52G000050ZFNoSAO/intercom-display-article-crash-on-ios
+  static Future<void> displayArticle(String articleId) async {
+    return IntercomFlutterPlatform.instance.displayArticle(articleId);
+  }
+
+  /// To display a Carousel, pass in a [carouselId] from your Intercom workspace.
+  ///
+  /// A carousel must be ‘live’ to be used in this feature.
+  /// If it is in a draft or paused state,
+  /// end-users will see an error if the app tries to open the content.
+  static Future<void> displayCarousel(String carouselId) async {
+    return IntercomFlutterPlatform.instance.displayCarousel(carouselId);
+  }
 }