+++ intercom_flutter/CHANGELOG.md
@@ -1,3 +1,15 @@
+# Changelog
+
+## 4.0.0
+* Bump Intercom Android SDK version to 10.0.0
+* Bump Intercom iOS SDK version to 10.0.0
+* Added support for displayCarousel
+* Added support for displayArticle
+    - Note: Intercom iOS SDK has an issue with displayArticle if your Intercom account does have that feature enabled. It crashes the app. The bug is already reported at https://forum.intercom.com/s/question/0D52G000050ZFNoSAO/intercom-display-article-crash-on-ios. As per the conversation with Intercom support, they are working on the issue. The fix may take some time.
+* Internal Changes:
+    - used `hideIntercom()` as `hideMessenger()` is deprecated and removed in Intercom SDK 10.0.0
+    - Android - updated gradle version and dependencies.
+
 ## 3.2.1
 * Fix `application has not been initialized` crash on Android when calling from background isolate.